Explorer: Split out `SolBalance` component from utils (#28072)
Explorer: Split out SolBalance component from utils
