+++ explorer/src/components/block/BlockRewardsCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { SolBalance } from "utils";
+import { SolBalance } from "components/common/SolBalance";
 import { PublicKey, VersionedBlockResponse } from "@solana/web3.js";
 import { Address } from "components/common/Address";
