+++ explorer/src/components/common/BalanceDelta.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { BigNumber } from "bignumber.js";
-import { SolBalance } from "utils";
+import { SolBalance } from "components/common/SolBalance";
 
 export function BalanceDelta({
   delta,
