+++ explorer/src/components/common/SolBalance.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { lamportsToSolString } from "utils";
+
+export function SolBalance({
+  lamports,
+  maximumFractionDigits = 9,
+}: {
+  lamports: number | bigint;
+  maximumFractionDigits?: number;
+}) {
+  return (
+    <span>
+      ◎
+      <span className="font-monospace">
+        {lamportsToSolString(lamports, maximumFractionDigits)}
+      </span>
+    </span>
+  );
+}
