+++ explorer/src/utils/index.ts
@@ -86,23 +86,6 @@ export function numberWithSeparator(s: string) {
   return s.replace(/\B(?=(\d{3})+(?!\d))/g, ",");
 }
 
-export function SolBalance({
-  lamports,
-  maximumFractionDigits = 9,
-}: {
-  lamports: number | bigint;
-  maximumFractionDigits?: number;
-}) {
-  return (
-    <span>
-      ◎
-      <span className="font-monospace">
-        {lamportsToSolString(lamports, maximumFractionDigits)}
-      </span>
-    </span>
-  );
-}
-
 const HUMANIZER = new HumanizeDuration(new HumanizeDurationLanguage());
 HUMANIZER.setOptions({
   language: "short",
